requirements.txt
tensorflow>=2.13
numpy
pandas
scikit-learn
matplotlib
opencv-python
Pillow
flask
werkzeug
tqdm


